dailybuild: boostをincludeすることによるwarning（マクロ再定義）を抑止
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7253 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHIKEngine.cpp
@@ -296,7 +296,7 @@ void PHIKEngine::Step() {
 	// <!!>
 	// ApplyExactState(/* reverse = */ true);
 
-	for (int i=0; i<numIter; ++i) {
+	for (size_t i=0; i<numIter; ++i) {
 		CalcJacobian();
 		IK();
 		Limit();
